add class Protein
